PIDCon.cpp
Might contain some errors. Please let me know some comments about the script so that I can keep updating the script accordingly. The script contains a flow of the different test cases  in which I attempt to tune the gains. I am updating the code to interface the PID gains with the on and off time of the valves.